refactor(core,common): rename ShadowEvent to SentinelEvent; extract HTTP middleware
Rename zeph_core::agent::shadow_sentinel::ShadowEvent to SentinelEvent across
all usage sites (shadow_sentinel.rs, builder.rs, runner.rs) to eliminate the
naming collision with zeph_sanitizer::ShadowMemory::ShadowEvent. ShadowEventStore
and ShadowEventRow are unambiguous and retained. A doc note is added to
ShadowMemory confirming it is not wired into the agent tool executor.

Extract AuthConfig, AuthIdentity, Cidr, RateLimitState, auth_middleware, and
rate_limit_middleware from zeph-gateway and zeph-a2a into a new
zeph-common::http_middleware module behind the http-middleware feature gate.
Gateway's pre-hashing approach (Option<blake3::Hash> at construction time,
subtle::ConstantTimeEq on 32-byte digests) is the canonical implementation.
Rate-limit uses gateway's inline GC and CIDR + XFF rightmost-untrusted IP
extraction. Both routers now import from the shared module.

Closes #4379, #4387.
