+++ crates/zeph-a2a/Cargo.toml
@@ -14,12 +14,11 @@ readme = "README.md"
 
 [features]
 ibct = ["dep:hmac", "dep:sha2"]
-server = ["dep:axum", "dep:blake3", "dep:subtle", "dep:tower", "dep:tower-http"]
+server = ["dep:axum", "dep:tower", "dep:tower-http", "zeph-common/http-middleware"]
 
 [dependencies]
 axum = { workspace = true, optional = true }
 base64.workspace = true
-blake3 = { workspace = true, optional = true }
 eventsource-stream.workspace = true
 futures.workspace = true
 futures-core.workspace = true
@@ -29,7 +28,6 @@ reqwest = { workspace = true, features = ["json", "stream"] }
 serde = { workspace = true, features = ["derive"] }
 serde_json.workspace = true
 sha2 = { workspace = true, optional = true }
-subtle = { workspace = true, optional = true }
 thiserror.workspace = true
 tokio = { workspace = true, features = ["net", "sync"] }
 tokio-stream.workspace = true
