+++ crates/zeph-common/Cargo.toml
@@ -13,6 +13,7 @@ description = "Shared utility functions and security primitives for Zeph crates"
 readme = "README.md"
 
 [features]
+http-middleware = ["dep:axum", "dep:subtle"]
 jsonschema = ["dep:schemars"]
 treesitter = [
     "dep:tree-sitter",
@@ -24,12 +25,14 @@ treesitter = [
 ]
 
 [dependencies]
+axum = { workspace = true, optional = true }
 blake3.workspace = true
 cpu-time.workspace = true
 metrics.workspace = true
 parking_lot.workspace = true
 schemars = { workspace = true, optional = true }
 serde = { workspace = true, features = ["derive", "rc"] }
+subtle = { workspace = true, optional = true }
 serde_json.workspace = true
 thiserror.workspace = true
 tokio = { workspace = true, features = ["macros", "rt", "rt-multi-thread", "sync", "time"] }
